dev-requirements.txt
black==22.3.0
pytest==7.1.2
pytest-cov==3.0.0
pydotenv==0.1.9
